Improved healthcheck and code clearing
