+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/sidecore/storage/RedisInfoParser.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.netflix.dynomitemanager.sidecore.utils;
+package com.netflix.dynomitemanager.sidecore.storage;
 
 import java.io.BufferedReader;
 import java.io.File;
