+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/sidecore/utils/ProcessMonitorTask.java
@@ -21,6 +21,7 @@ import com.netflix.dynomitemanager.sidecore.scheduler.SimpleTimer;
 import com.netflix.dynomitemanager.sidecore.scheduler.Task;
 import com.netflix.dynomitemanager.sidecore.scheduler.TaskTimer;
 import com.netflix.dynomitemanager.sidecore.storage.IStorageProxy;
+import com.netflix.dynomitemanager.sidecore.storage.JedisUtils;
 
 import org.quartz.StatefulJob;
 import org.slf4j.Logger;
